pyproject.toml
[tool.poetry]
name = "wave-poc"
version = "0.1.0"
description = ""
authors = ["Kanisk Chakraborty <57884127+chakrakan@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "wave_poc"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
